preventing category and tag to be deleted if has post

+++ app/Http/Controllers/CategoriesController.php
@@ -94,6 +94,13 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category cannot be deleted as it has posts.');    
+
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category has been deleted successfully.');
