+++ app/Http/Controllers/TagController.php
@@ -94,6 +94,13 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        if ($tag->posts->count() > 0) {
+            session()->flash('error', 'Tag cannot be deleted as it has posts.');    
+
+            return redirect()->back();
+        }
+
         $tag->delete();
 
         session()->flash('success', 'Tag has been deleted successfully.');
